clean code of get reservations
